Refactor main into modular utilities
